Allow fractional per-case latency_ms
External agent-connection tests self-report metrics.latency_ms verbatim,
which can be a float (e.g. 1955.7) — only calibrate's internal-model path
rounds to an int. The TestCaseResult.latency_ms field was Optional[int], so
a fractional value raised a pydantic ValidationError on the run-status
response. Widen to Optional[float] (still accepts ints).

+++ src/routers/agent_tests.py
@@ -274,8 +274,10 @@ class TestCaseResult(BaseModel):
     judge_results: Optional[List[JudgeResult]] = None
     # Response-generation latency for this case in milliseconds (the agent under
     # test, not the judge). Present only for live runs; None for in-progress rows
-    # and eval-only runs (no inference happened).
-    latency_ms: Optional[int] = None
+    # and eval-only runs (no inference happened). Float, not int: calibrate rounds
+    # internal-model latency to a whole number, but external agent-connection tests
+    # self-report `metrics.latency_ms` verbatim and may send a fractional value.
+    latency_ms: Optional[float] = None
     # Per-case cost in USD, lifted from calibrate's nested `output.cost` (NOT a
     # top-level sibling of latency_ms — cost lives inside `output`). None when the
     # provider/agent reports none (e.g. --provider openai, or no `metrics.cost`).
